Use composition to allow reuse of WorkflowInvocationState's state logic
