+++ client/src/stores/invocationStore.ts
@@ -4,8 +4,8 @@ import {
     fetchInvocationDetails,
     fetchInvocationJobsSummary,
     fetchInvocationStep,
-    type WorkflowInvocation,
     type InvocationJobsSummary,
+    type WorkflowInvocation,
     type WorkflowInvocationStep,
 } from "@/api/invocations";
 import { useKeyedCache } from "@/composables/keyedCache";
